checkout: Recurring order UI updated

+++ VirtoCommerce.Storefront/App_Data/Themes/default/assets/js/checkout/checkout.js
@@ -51,8 +51,8 @@ angular.module(moduleName, ['credit-cards', 'angular.filter'])
                     }
                     $scope.checkout.billingAddressEqualsShipping = cart.hasPhysicalProducts && !angular.isObject($scope.checkout.payment.billingAddress);
 
-                    $scope.checkout.canSetCartRecurring = !cart.paymentPlan && _.all(cart.items, function (x) { return !x.isReccuring });
-                    $scope.paymentPlan = cart.paymentPlan || ($scope.checkout.canSetCartRecurring && { intervalCount: 1, interval: 'months' });
+                    $scope.checkout.canCartBeRecurring = _.all(cart.items, function (x) { return !x.isReccuring });
+                    $scope.checkout.paymentPlan = cart.paymentPlan && _.findWhere($scope.paymentPlans, { intervalCount: cart.paymentPlan.intervalCount, interval: cart.paymentPlan.interval });
                 }
                 $scope.validateCheckout($scope.checkout);
                 return cart;
@@ -153,16 +153,27 @@ angular.module(moduleName, ['credit-cards', 'angular.filter'])
             });
         };
 
+        $scope.paymentPlans = [// { intervalCount: 1, interval: 'days' },
+                               { intervalCount: 1, interval: 'weeks', name: '1 week' },
+                               { intervalCount: 2, interval: 'weeks', name: '2 weeks' },
+                               { intervalCount: 1, interval: 'months', name: '1 month' },
+                               { intervalCount: 2, interval: 'months', name: '2 months' },
+                               { intervalCount: 3, interval: 'months', name: '3 months' },
+                               { intervalCount: 4, interval: 'months', name: '4 months' },
+                               { intervalCount: 6, interval: 'months', name: '6 months' }];
+
         $scope.savePaymentPlan = function () {
             return wrapLoading(function () {
-                return cartService.addOrUpdatePaymentPlan($scope.paymentPlan).then($scope.reloadCart);
+                return cartService.addOrUpdatePaymentPlan($scope.checkout.paymentPlan).then($scope.reloadCart);
             });
         };
 
-        $scope.deletePaymentPlan = function () {
-            return wrapLoading(function () {
-                return cartService.removePaymentPlan().then($scope.reloadCart);
-            });
+        $scope.isRecurringChanged = function (isRecurring) {
+            if (!isRecurring && $scope.checkout.paymentPlan) {
+                return wrapLoading(function () {
+                    return cartService.removePaymentPlan().then($scope.reloadCart);
+                });
+            }
         };
 
         function updatePayment(payment) {
